Lab6/requirements.txt
Pillow==9.2.0
tk==0.1.0